rename: prismaModule, prismaService 폴더 이동 #16
