+++ src/domain/review/repository/short-review.repository.ts
@@ -3,13 +3,13 @@ import {
   InternalServerErrorException,
   NotFoundException,
 } from '@nestjs/common';
-import { PrismaService } from '../../../global/config/prisma/prisma.service';
 import {
   IReviewQuery,
   IShortReview,
   SortCondition,
 } from '../reviews.interface';
 import { AttractionElement, Prisma, ReviewType } from '@prisma/client';
+import { PrismaService } from '../../../global/database/prisma/prisma.service';
 
 @Injectable()
 export class ShortReviewRepository {
