--- src/global/config/prisma/prisma.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PrismaService } from './prisma.service';
-import { ConfigService } from '@nestjs/config';
-
-@Module({
-  providers: [PrismaService, ConfigService],
-  exports: [PrismaService],
-})
-export class PrismaModule {}
