--- src/global/config/prisma/prisma.service.ts
@@ -1,48 +1,0 @@
-import { Injectable, Logger, OnModuleInit } from '@nestjs/common';
-import { Prisma, PrismaClient } from '@prisma/client';
-import { ConfigService } from '@nestjs/config';
-
-@Injectable()
-export class PrismaService
-  extends PrismaClient<Prisma.PrismaClientOptions, 'query' | 'error'>
-  implements OnModuleInit
-{
-  private readonly logger = new Logger(PrismaService.name);
-
-  constructor(private readonly configService: ConfigService) {
-    super({
-      log: [
-        {
-          emit: 'event',
-          level: 'query',
-        },
-        {
-          emit: 'event',
-          level: 'error',
-        },
-        {
-          emit: 'stdout',
-          level: 'info',
-        },
-        {
-          emit: 'stdout',
-          level: 'warn',
-        },
-      ],
-    });
-  }
-
-  async onModuleInit(): Promise<void> {
-    if (this.configService.get('NODE_ENV') == 'dev') {
-      this.$on('query', (event) => {
-        this.logger.verbose(event.query, event.duration);
-      });
-    }
-
-    this.$on('error', (event) => {
-      this.logger.verbose(event.target);
-    });
-
-    await this.$connect();
-  }
-}
