+++ src/global/database/prisma/prisma.module.ts
@@ -1,9 +1,10 @@
 import { Global, Module } from '@nestjs/common';
-import { PrismaService } from 'src/global/database/prisma/prisma.service';
+import { PrismaService } from './prisma.service';
+import { ConfigService } from '@nestjs/config';
 
 @Global()
 @Module({
-  providers: [PrismaService],
+  providers: [PrismaService, ConfigService],
   exports: [PrismaService],
 })
 export class PrismaModule {}
